Added new widget: wm.DesignableDialog
This widget allows the user to drag and drop widgets into a dialog which is then written to widgets.js.  Widgets in wm.DesignableDialog are owned not by the Dialog, but instead by the page. 

This means that the widgets event handlers and binding can be to other page level elements, and other page level elements can bind to widgets in the dialog.

Known issues:
1. Have had to disable making this an app level dialog, must be page level to be designable.
2. Widgets in model show up fully expanded, making the page section of the model harder to use

git-svn-id: https://src.springsource.com/svn/wm_svn/eng3/trunk@29813 bab4cc61-abf4-0310-9601-8b162d8b0fb1