Adds a scrollToRow method and scrollToSelection property
DojoGrid automatically scrolls to the selection when the developer calls select.  This is a half-assed version of that for wm.List.

Issue: it can only scroll to the specified item by rendering all hidden rows which is very bad for large datasets.  There are better ways to do this, but this is a last minute hack that will be off by default.