Insures nonvisual components are written before visual Issues: WM-4665
This issue became serious once wm.LiveVariable could write its LiveView to widgets.js.  This change requires LiveViews to be generated between LiveVariable init and postInit.  A notification of change is fired during init at which time it lacks a LiveView, and subscribed grids see a change in dataSet but no LiveView to indicate what fields there are (result: no fields and so no columns)