Adds support for theme paths and for "project" module 1. Themes can now come from anywhere.  Instead of using a theme name, the theme property is now a theme path expressed using dojo module syntax "lib.wm.base.widget.theme.wm_default"
Why?  This allows themes to be stored in the project resources
folder, in a Composite/custom widget folder, or anywhere else.

2. Changed loader.js, runtimeLoader.js and project.js to
register the module "project" and for dojo to treat it as a "fixed"
path rather than something to be adjusted using the dojo.baseUrl.

Why? We can now set a project theme to be "project.resources.mytheme".
We can now have dojo.require("project.resources.javascript.mylib");

3. Adds app.css to Composite css in Publisher.js

4. Adds support in propertyEdit for a wm.prop.Select editor to
use restrictValues=false.  This is now used in Application_design.js